requirements.txt
quixstreams
requests
pandas
duckdb
streamlit
python-dotenv
pytrends